fix rebase

+++ packages/flutter/lib/src/widgets/navigator.dart
@@ -3470,7 +3470,7 @@ class NavigatorState extends State<Navigator> with TickerProviderStateMixin, Res
   @override
   void didChangeDependencies() {
     super.didChangeDependencies();
-    _viewId = View.of(context).viewId as int;
+    _viewId = View.of(context).viewId;
     _updateHeroController(HeroControllerScope.maybeOf(context));
     for (final _RouteEntry entry in _history) {
       entry.route.changedExternalState();
